Test different methods of scraping playlists
scrape.py is the first version, it only scrapes stuff from a non-dynamic
page, so it doesn't work for spotify. test.py is improved, as it can
read from dynamic pages and can read spotify playlists, but it can't
scroll, so it only reads from the first few songs of the playlist.
test2.py is an improvement, as it can scroll, but it scrolls too fast
and misses a bunch of songs from the playlist.